I began work implementing an input manager. It reads input bindings from a JSON file using JSONFile and uses Boost.Signals to allow clients to register themselves as listeners to input events. Events may have multiple listeners.
In order to make the same manager viable for use across the entire game, commands are categorized according to "context." In the JSON, a context is just an object filled with command bindings. For example, there may be the object "Menu," one of whose members is "Quit" : "Space". The "Quit" event, triggered by the spacebar, would only actually be fired if the current context of the manager is set to "Menu".

To subscribe a function to listen to the above event, the syntax that the hooking function would require would be "Menu.Quit:Down" or "Menu.Quit:Up". The abstract syntax is "Context.CommandName:Direction". Setting Direction to "Down" causes it to react with the associated key is pressed down; setting it to "Up" causes it to react with the associated key is released.

There are a few issues right now. The system is just now implemented and not interfaced with the game state system fully. There appear to be some issues actually reading in the bindings from the JSON file, but I tested it by manually setting the key that is associated with some sample input events, and it works fine. Finally, I'm not sure if it's my fault or Gosu's, but it doesn't appear to behave quite directly when trying to detect key releases. It seems to fire the event when the key is pressed down a second time rather than when it is released the first time.