Inventory + Items + Weapons
Items and Weapons can exist in the scene and also in the players inventory.
This single class system allows us to have persistant objects during runtime
and not deal with translation of scene object to inventory object.  

When a player grabs an object, it is checked via a bitmask to see if it is an inventory
item or physics object.  When it is an inventory item, it is passed onto the inventory
which will place it in either the weapon list or item list.  These lists will be used for 
inventory HUDs and menus and also passing messages to the currently held weapons.

Inventory items have icons which are shown in the inventory HUD.  Weapons have 
sprites which are drawn with the player to show them being carried in hand.  

More to come as I flesh it out.