Integrated InputManager into MainMenu
This required providing and interface on the MUGE class, which required making the typedefs public members of the class.

Also changed GameState to keep a pointer to the engine, and only require sending a pointer to MUGE in the init.  Further calls just use the pointer when necessary.

Turns out Gosu does not send mouse button events through the button down callbacks, so we'll need to continue using mouse interaction through the input() accessor.  This isn't really a bad thing, as most mouse interaction is continuous.